Fixes back button behaviour for content pages in Learn

+++ kolibri/plugins/learn/assets/src/views/TopicsContentPage.vue
@@ -1,7 +1,7 @@
 <template>
 
   <div
-    v-if="!loading"
+    v-if="!loading && content"
     ref="mainWrapper"
     class="main-wrapper"
   >
@@ -148,10 +148,10 @@
   import AppError from 'kolibri-common/components/AppError';
   import GlobalSnackbar from 'kolibri-common/components/GlobalSnackbar';
   import SkipNavigationLink from '../../../../../../kolibri/core/assets/src/views/SkipNavigationLink';
+  import useContentLink from '../composables/useContentLink';
   import useCoreLearn from '../composables/useCoreLearn';
   import useContentNodeProgress from '../composables/useContentNodeProgress';
   import useLearnerResources from '../composables/useLearnerResources';
-  import { PageNames } from '../constants';
   import SidePanelModal from './SidePanelModal';
   import LearningActivityChip from './LearningActivityChip';
   import LessonResourceViewer from './classes/LessonResourceViewer';
@@ -202,12 +202,14 @@
         contentNodeProgressMap,
       } = useContentNodeProgress();
       const { fetchLesson } = useLearnerResources();
+      const { back } = useContentLink();
       return {
         canDownload,
         contentNodeProgressMap,
         fetchContentNodeProgress,
         fetchContentNodeTreeProgress,
         fetchLesson,
+        back,
       };
     },
     props: {
@@ -247,13 +249,13 @@
       }),
       ...mapState('topicsTree', ['content']),
       ...mapState('topicsTree', {
-        isCoachContent: state => (state.content.coach_content ? 1 : 0),
+        isCoachContent: state => (state.content && state.content.coach_content ? 1 : 0),
       }),
       practiceQuiz() {
         return get(this, ['content', 'options', 'modality']) === Modalities.QUIZ;
       },
       contentProgress() {
-        return this.contentNodeProgressMap[this.content.content_id];
+        return this.contentNodeProgressMap[this.content && this.content.content_id];
       },
       notAuthorized() {
         // catch "not authorized" error, display AuthMessage
@@ -288,35 +290,6 @@
       },
       timeSpent() {
         return this.contentPageMounted ? this.$refs.contentPage.time_spent : 0;
-      },
-      back() {
-        if (!this.$route) {
-          return null;
-        }
-        const query = { ...this.$route.query };
-        const lastPage = (this.$route.query || {}).last;
-        delete query.last;
-        delete query.topicId;
-        // returning to a topic page requires an id
-        if (lastPage === PageNames.TOPICS_TOPIC_SEARCH || lastPage === PageNames.TOPICS_TOPIC) {
-          const lastId = this.$route.query.topicId
-            ? this.$route.query.topicId
-            : this.content.parent;
-          // Need to guard for parent being non-empty to avoid console errors
-          return this.$router.getRoute(
-            lastPage,
-            {
-              id: lastId,
-            },
-            query
-          );
-        } else if (lastPage === PageNames.LIBRARY) {
-          return this.$router.getRoute(lastPage, {}, query);
-        } else if (lastPage) {
-          return this.$router.getRoute(lastPage, query);
-        } else {
-          return this.$router.getRoute(PageNames.HOME);
-        }
       },
     },
     watch: {
@@ -409,6 +382,9 @@
        * is logged in
        */
       fetchSiblings() {
+        if (!this.content) {
+          return Promise.resolve();
+        }
         // Fetch the next content
         const nextPromise = ContentNodeResource.fetchNextContent(this.content.parent, {
           topicOnly: true,
@@ -448,6 +424,9 @@
         this.sidePanelContent = this.content;
       },
       toggleBookmark() {
+        if (!this.content) {
+          return;
+        }
         if (this.bookmark) {
           client({
             method: 'delete',
